Outsourced logger_init to its own module and split up components
