+++ src/main.rs
@@ -5,7 +5,7 @@ mod kaenguru;
 mod xd;
 mod config;
 mod replies;
-mod logger_init;
+mod logger;
 
 use std::{
     borrow::Borrow,
@@ -105,7 +105,7 @@ async fn start_kg() {
 
 /// Main entry point to this program
 fn main() {
-    logger_init::init();
+    logger::init();
 
     println!("[MAIN]:\tStarting \"Känguru Rechenkencht\" and \"XD-Bot\"...");
     // Use tokio to run multiple bots at the same time
